Forgot to add a short-circuit guard

+++ muxer/meshuggahme_muxer.py
@@ -11,7 +11,7 @@ app = Flask(app_name)
 video_url_prefix = os.environ.get('MUSHUGGAHME_OUTPUT_URL', 'http://localhost:5000/video')
 
 pidfile = os.environ.get('MESHUGGAHME_PIDFILE', None)
-if os.path.exists(os.path.dirname(pidfile)):
+if pidfile and os.path.exists(os.path.dirname(pidfile)):
     with open(pidfile, 'w') as f:
         f.write(os.getpid())
